+++ docs/CHANGES.rst
@@ -6,8 +6,6 @@ There's a frood who really knows where his towel is
 
 - The WCGA 1.0 and WCGA 2.0 specifications and validators require some
   changes for acessibility issues:
-  * You can't use fixed value in width and height (this will be overriden
-    with css in your site) [rodfersou].
   * You need a title attribute in the iframe to make some screen readers
     for blind people work [rodfersou].
   * You need to add a link into the iframe tag to point to the content,
@@ -15,6 +13,9 @@ There's a frood who really knows where his towel is
     it behaves as alternative content [rodfersou].
 * The W3C Validator suggest that the attribute frameborder is obsolete,
   so I take it out if I find this attribute into the iframe element
+  [rodfersou].
+* Also @hvelarde pointed me that allowfullscreen, mozallowfullscreen and
+  webkitallowfullscreen are obsolete too, so I took them out too
   [rodfersou].